requirements.txt
beautifulsoup4==4.12.3
pandas==2.0.3
python-dotenv==1.0.1
Requests==2.31.0
